feat: update referenced to moved layout components
