+++ src/views/PortfolioItem.vue
@@ -258,7 +258,7 @@
 <script>
 import Vue             from 'vue'
 import { mapGetters }  from 'vuex'
-import NavBar          from '../components/Nav'
+import NavBar          from '../layout-components/Nav.vue'
 import VueTypedJs      from 'vue-typed-js'
 import HighlightBanner from "./HighlightBanner";
